+++ rdf_utils.h
@@ -32,6 +32,7 @@ extern char *QuoteRDFLiteral(const char *input);
 extern char *EscapeSPARQLStringContent(const char *str, bool escape_backslash);
 extern char *cstring_to_rdfliteral(char *input);
 extern char *EscapeSPARQLLiteral(const char *input);
+extern char *XPathReplacementToPG(const char *replacement);
 extern char *ExpandDatatypePrefix(char *str);
 extern char *unescape_unicode(const char *input);